swift_hire/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic[email]==2.10.4
python-dotenv==1.0.1
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0